[WIP/debug] support non-singlet unrestricted spin cases
This commit is working in the sense that the correct mathematics are in
place. However, we are now running into an issue with respect to
mismatching numbers of spin orbitals per spin species when the number of
active basis functions is chosen too little. For this, there are two
possible paths forward:

1. enforce a matching number of spin orbitals per spin species during
   the concentric localitzation technique
2. extend Qiskit Nature to support mismatched spin register sizes